PlatformPayloadFeaturePkg: Support board and common features
As part of the UPL preparation work, we are migrating to a new design
where downstream features are maintained in the platform feature FV.
This allows us to switch to, or closely track, upstream edk2.

We design for modularity by enabling a set of of common features to all
downstream platforms, and optional board-specific features. Define the
`BOARD` build definition on the command line to compile its features.

By minimising the patches to upstream code and using new directories, we
can rebase quickly to follow changes upstream.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>